Move UserRole to account package
