--- backend/src/main/java/com/karankumar/bookproject/backend/model/UserRole.java
@@ -1,17 +1,0 @@
-package com.karankumar.bookproject.backend.model;
-
-public enum UserRole {
-    ADMIN("ADMIN"),
-    USER("USER");
-
-    private final String role;
-
-    UserRole(String role) {
-        this.role = role;
-    }
-
-    @Override
-    public String toString() {
-        return role;
-    }
-}
