+++ backend/src/main/java/com/karankumar/bookproject/backend/model/account/UserRole.java
@@ -1,0 +1,31 @@
+/*
+ * The book project lets a user keep track of different books they would like to read, are currently
+ * reading, have read or did not finish.
+ * Copyright (C) 2021  Karan Kumar
+ * This program is free software: you can redistribute it and/or modify it under the terms of the
+ * GNU General Public License as published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY
+ * WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
+ * PURPOSE.  See the GNU General Public License for more details.
+ * You should have received a copy of the GNU General Public License along with this program.
+ * If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package com.karankumar.bookproject.backend.model.account;
+
+public enum UserRole {
+    ADMIN("ADMIN"),
+    USER("USER");
+
+    private final String role;
+
+    UserRole(String role) {
+        this.role = role;
+    }
+
+    @Override
+    public String toString() {
+        return role;
+    }
+}
