+++ backend/src/main/java/com/karankumar/bookproject/backend/service/UserService.java
@@ -17,7 +17,7 @@
 
 package com.karankumar.bookproject.backend.service;
 
-import com.karankumar.bookproject.backend.model.UserRole;
+import com.karankumar.bookproject.backend.model.account.UserRole;
 import com.karankumar.bookproject.backend.model.account.Role;
 import com.karankumar.bookproject.backend.model.account.User;
 import com.karankumar.bookproject.backend.repository.RoleRepository;
