+++ backend/src/test/java/com/karankumar/bookproject/backend/service/UserServiceTest.java
@@ -18,7 +18,7 @@
 package com.karankumar.bookproject.backend.service;
 
 import com.karankumar.bookproject.annotations.IntegrationTest;
-import com.karankumar.bookproject.backend.model.UserRole;
+import com.karankumar.bookproject.backend.model.account.UserRole;
 import com.karankumar.bookproject.backend.model.account.Role;
 import com.karankumar.bookproject.backend.model.account.User;
 import com.karankumar.bookproject.backend.repository.RoleRepository;
